Move more symbols into Foundation
Symbol, Cursor, SortField, OverlapStrategy, and Side
