+++ hvega/src/Graphics/Vega/VegaLite/Foundation.hs
@@ -34,6 +34,15 @@ module Graphics.Vega.VegaLite.Foundation
 
        , Scale(..)
 
+       , SortField(..)
+
+       , Cursor(..)
+
+       , OverlapStrategy(..)
+       , Side(..)
+
+       , Symbol(..)
+
        -- not for external export
        , fontWeightSpec
        , measurementLabel
@@ -46,20 +55,35 @@ module Graphics.Vega.VegaLite.Foundation
        , strokeJoinLabel
        , scaleLabel
        , positionLabel
+       , sortFieldSpec
+       , cursorLabel
+       , overlapStrategyLabel
+       , sideLabel
+       , symbolLabel
 
+       , field_
+       , order_
        )
     where
 
 import qualified Data.Text as T
 
-import Data.Aeson (toJSON)
+import Data.Aeson ((.=), object, toJSON)
 
 
 -- added in base 4.8.0.0 / ghc 7.10.1
 import Numeric.Natural (Natural)
 
 
-import Graphics.Vega.VegaLite.Specification (VLSpec)
+import Graphics.Vega.VegaLite.Specification (VLSpec, LabelledSpec)
+
+
+field_ :: T.Text -> LabelledSpec
+field_ f = "field" .= f
+
+-- could restrict to ascending/descending
+order_ :: T.Text -> LabelledSpec
+order_ o = "order" .= o
 
 
 {-|
@@ -504,3 +528,213 @@ scaleLabel ScBinOrdinal = "bin-ordinal"
 scaleLabel ScQuantile = "quantile"
 scaleLabel ScQuantize = "quantize"
 scaleLabel ScThreshold = "threshold"
+
+
+-- | How should the field be sorted when performing a window transform.
+--
+--   @since 0.4.00
+
+data SortField
+    = WAscending T.Text
+    -- ^ Sort the field into ascending order.
+    | WDescending T.Text
+    -- ^ Sort the field into descending order.
+
+
+sortFieldSpec :: SortField -> VLSpec
+sortFieldSpec (WAscending f) = object [field_ f, order_ "ascending"]
+sortFieldSpec (WDescending f) = object [field_ f, order_ "descending"]
+
+
+{-|
+
+Represents the type of cursor to display. For an explanation of each type,
+see the
+<https://developer.mozilla.org/en-US/docs/Web/CSS/cursor#Keyword%20values CSS documentation>.
+
+-}
+data Cursor
+    = CAuto
+    | CDefault
+    | CNone
+    | CContextMenu
+    | CHelp
+    | CPointer
+    | CProgress
+    | CWait
+    | CCell
+    | CCrosshair
+    | CText
+    | CVerticalText
+    | CAlias
+    | CCopy
+    | CMove
+    | CNoDrop
+    | CNotAllowed
+    | CAllScroll
+    | CColResize
+    | CRowResize
+    | CNResize
+    | CEResize
+    | CSResize
+    | CWResize
+    | CNEResize
+    | CNWResize
+    | CSEResize
+    | CSWResize
+    | CEWResize
+    | CNSResize
+    | CNESWResize
+    | CNWSEResize
+    | CZoomIn
+    | CZoomOut
+    | CGrab
+    | CGrabbing
+
+
+cursorLabel :: Cursor -> T.Text
+cursorLabel CAuto = "auto"
+cursorLabel CDefault = "default"
+cursorLabel CNone = "none"
+cursorLabel CContextMenu = "context-menu"
+cursorLabel CHelp = "help"
+cursorLabel CPointer = "pointer"
+cursorLabel CProgress = "progress"
+cursorLabel CWait = "wait"
+cursorLabel CCell = "cell"
+cursorLabel CCrosshair = "crosshair"
+cursorLabel CText = "text"
+cursorLabel CVerticalText = "vertical-text"
+cursorLabel CAlias = "alias"
+cursorLabel CCopy = "copy"
+cursorLabel CMove = "move"
+cursorLabel CNoDrop = "no-drop"
+cursorLabel CNotAllowed = "not-allowed"
+cursorLabel CAllScroll = "all-scroll"
+cursorLabel CColResize = "col-resize"
+cursorLabel CRowResize = "row-resize"
+cursorLabel CNResize = "n-resize"
+cursorLabel CEResize = "e-resize"
+cursorLabel CSResize = "s-resize"
+cursorLabel CWResize = "w-resize"
+cursorLabel CNEResize = "ne-resize"
+cursorLabel CNWResize = "nw-resize"
+cursorLabel CSEResize = "se-resize"
+cursorLabel CSWResize = "sw-resize"
+cursorLabel CEWResize = "ew-resize"
+cursorLabel CNSResize = "ns-resize"
+cursorLabel CNESWResize = "nesw-resize"
+cursorLabel CNWSEResize = "nwse-resize"
+cursorLabel CZoomIn = "zoom-in"
+cursorLabel CZoomOut = "zoom-out"
+cursorLabel CGrab = "grab"
+cursorLabel CGrabbing = "grabbing"
+
+
+{-|
+
+Type of overlap strategy to be applied when there is not space to show all items
+on an axis. See the
+<https://vega.github.io/vega-lite/docs/axis.html#labels Vega-Lite documentation>
+for more details.
+-}
+
+data OverlapStrategy
+    = ONone
+      -- ^ No overlap strategy to be applied when there is not space to show all items
+      --   on an axis.
+    | OParity
+      -- ^ Give all items equal weight in overlap strategy to be applied when there is
+      --   not space to show them all on an axis.
+    | OGreedy
+      -- ^ Greedy overlap strategy to be applied when there is not space to show all
+      --   items on an axis.
+
+overlapStrategyLabel :: OverlapStrategy -> T.Text
+overlapStrategyLabel ONone = "false"
+overlapStrategyLabel OParity = "parity"
+overlapStrategyLabel OGreedy = "greedy"
+
+
+-- | Represents one side of a rectangular space.
+
+data Side
+    = STop
+    | SBottom
+    | SLeft
+    | SRight
+
+
+sideLabel :: Side -> T.Text
+sideLabel STop = "top"
+sideLabel SBottom = "bottom"
+sideLabel SLeft = "left"
+sideLabel SRight = "right"
+
+
+-- | Identifies the type of symbol used with the 'Graphics.Vega.VegaLite.Point' mark type.
+--   It is used with 'Graphics.Vega.VegaLite.MShape', 'Graphics.Vega.VegaLite.LeSymbolType', and 'Graphics.Vega.VegaLite.LSymbolType'.
+--
+--   In version @0.4.0.0@ all constructors were changed to start
+--   with @Sym@.
+--
+data Symbol
+    = SymCircle
+      -- ^ Specify a circular symbol for a shape mark.
+    | SymSquare
+      -- ^ Specify a square symbol for a shape mark.
+    | SymCross
+      -- ^ Specify a cross symbol for a shape mark.
+    | SymDiamond
+      -- ^ Specify a diamond symbol for a shape mark.
+    | SymTriangleUp
+      -- ^ Specify an upward-triangular symbol for a shape mark.
+    | SymTriangleDown
+      -- ^ Specify a downward-triangular symbol for a shape mark.
+    | SymTriangleRight
+      -- ^ Specify an right-facing triangular symbol for a shape mark.
+      --
+      --   @since 0.4.0.0
+    | SymTriangleLeft
+      -- ^ Specify an left-facing triangular symbol for a shape mark.
+      --
+      --   @since 0.4.0.0
+    | SymStroke
+      -- ^ The line symbol.
+      --
+      --  @since 0.4.0.0
+    | SymArrow
+      -- ^ Centered directional shape.
+      --
+      --  @since 0.4.0.0
+    | SymTriangle
+      -- ^ Centered directional shape. It is not clear what difference
+      --   this is to 'SymTriangleUp'.
+      --
+      --  @since 0.4.0.0
+    | SymWedge
+      -- ^ Centered directional shape.
+      --
+      --  @since 0.4.0.0
+    | SymPath T.Text
+      -- ^ A custom symbol shape as an
+      --   [SVG path description](https://developer.mozilla.org/en-US/docs/Web/SVG/Tutorial/Paths).
+      --
+      --   For correct sizing, the path should be defined within a square
+      --   bounding box, defined on an axis of -1 to 1 for both dimensions.
+
+
+symbolLabel :: Symbol -> T.Text
+symbolLabel SymCircle = "circle"
+symbolLabel SymSquare = "square"
+symbolLabel SymCross = "cross"
+symbolLabel SymDiamond = "diamond"
+symbolLabel SymTriangleUp = "triangle-up"
+symbolLabel SymTriangleDown = "triangle-down"
+symbolLabel SymTriangleRight = "triangle-right"
+symbolLabel SymTriangleLeft = "triangle-left"
+symbolLabel SymStroke = "stroke"
+symbolLabel SymArrow = "arrow"
+symbolLabel SymTriangle = "triangle"
+symbolLabel SymWedge = "wedge"
+symbolLabel (SymPath svgPath) = svgPath
